+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+- Fixed aliasing in Sinc interpolator when downsampling by adding automatic
+  anti-aliasing filter cutoff adjustment. The `Interpolator` trait now includes
+  a `set_bandwidth` method (with default no-op implementation) that is called
+  automatically by `Converter`.
 - Renamed `window-hanning` to `window-hann`
 - Made `IntoInterleavedSamples` and `IntoInterleavedSamplesIterator` stop
   yielding samples when the underlying signal gets exhausted. This is a breaking